Refactored test case
